// generated/src/app/api-models/id-card-narrative.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IdCardNarrative {
  seqNarrative: number;
  description: string;
  narrative: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  userDefined6: string;
  userDefined7: string;
  userDefined8: string;
  userDefined9: string;
  userDefined10: string;
  userDefined11: string;
  userDefined12: string;
  userDefined13: string;
  userDefined14: string;
  userDefined15: string;
  userDefined16: string;
  userDefined17: string;
  userDefined18: string;
  userDefined19: string;
  userDefined20: string;
}